DiscordSoundHub-1/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.14",
    "cffi>=1.17.1",
    "discord-py>=2.5.2",
    "email-validator>=2.2.0",
    "ffmpeg>=1.4",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydub>=0.25.1",
    "pynacl>=1.5.0",
    "python-dotenv>=1.1.0",
    "werkzeug>=3.1.3",
]
